% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vis3D.R
\name{renderVis3D}
\alias{renderVis3D}
\title{Widget render function for use in Shiny}
\usage{
renderVis3D(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

