% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vis3D.R
\name{vis3D}
\alias{vis3D}
\title{3D visualization}
\usage{
vis3D(data, style = "surface", width = NULL, height = NULL)
}
\arguments{
\item{data}{: data.frame. Needed at least columns "x", "y" and "z"
See \url{http://visjs.org/docs/graph3d.html#Data_Format}
\itemize{
 \item{"x"}{ : Location on the x-axis.}
 \item{"y"}{ : Location on the y-axis.}
 \item{"z"}{ : Location on the z-axis.}
 \item{"style"}{ : The data value, required for graph styles bar-color, bar-size, dot-color and dot-size. }
 \item{"filter"}{ : Filter values used for the animation. This column may have any type, such as a number, string, or Date.}
}}

\item{style}{: The style of the 3d graph. Available styles: bar, bar-color, bar-size, dot, dot-line, dot-color, dot-size, line, grid, or surface}
}
\description{
3D visualization using vis.js library
}
\examples{
nb <- 10
don <- expand.grid(x = 1:nb, y = 1:nb)
don$z <-rnorm(nb*nb)

vis3D(don)
vis3D(don, style = "dot-line")
vis3D(don, style = "grid")
vis3D(don, style = "bar")

# style (dot-color and dot-size)
nb <- 10
don <- expand.grid(x = 1:nb, y = 1:nb)
don$z <-rnorm(nb*nb)
don$style <-runif(nb*nb)

vis3D(don, style = "dot-size")
vis3D(don, style = "dot-color")

#tooltip
vis3D(don) \%>\% visOptions3D(tooltip = TRUE)

#custom tooltip
vis3D(don) \%>\% visOptions3D(tooltip = "function (point) {
 return 'value: <b>' + point.z + '</b>';}")

# customize label
vis3D(don) \%>\% visOptions3D(xValueLabel = "function (x) {return (x) + '\%'}")

#animation
nb <- 5
don <- expand.grid(x = 1:nb, y = 1:nb,
 filter = as.character(seq(as.Date("1990/1/1"), as.Date("1999/1/1"), "years")))
don$z <- rnorm(nrow(don))

vis3D(don)

#tooltip
}
\seealso{
\link{visOptions3D}
}

