% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vis3D.R
\name{vis3DOutput}
\alias{vis3DOutput}
\title{Widget output function for use in Shiny}
\usage{
vis3DOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

