% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visOptions.R
\name{visOptions3D}
\alias{visOptions3D}
\title{3D visualization general options}
\usage{
visOptions3D(graph, animationInterval = NULL, animationPreload = NULL,
  animationAutoStart = NULL, backgroundColor = NULL,
  cameraPosition = NULL, height = NULL, keepAspectRatio = NULL,
  showAnimationControls = NULL, showGrid = NULL, showPerspective = NULL,
  showShadow = NULL, style = NULL, tooltip = NULL, valueMax = NULL,
  valueMin = NULL, verticalRatio = NULL, width = NULL, xBarWidth = NULL,
  xCenter = NULL, xMax = NULL, xMin = NULL, xStep = NULL,
  xValueLabel = NULL, yBarWidth = NULL, yCenter = NULL, yMax = NULL,
  yMin = NULL, yStep = NULL, yValueLabel = NULL, zMin = NULL,
  zMax = NULL, zStep = NULL, zValueLabel = NULL, xLabel = NULL,
  yLabel = NULL, zLabel = NULL, filterLabel = NULL, legendLabel = NULL)
}
\arguments{
\item{animationInterval}{: number. Default to 1000. The animation interval in milliseconds. This determines how fast the animation runs.}

\item{animationPreload}{: boolean. Default to false. If false, the animation frames are loaded as soon as they are requested. if animationPreload is true, the graph will automatically load all frames in the background, resulting in a smoother animation as soon as all frames are loaded. The load progress is shown on screen.}

\item{animationAutoStart}{: boolean. Default to false. If true, the animation starts playing automatically after the graph is created.}

\item{backgroundColor}{: Just a String, or a named list. The background color for the main area of the chart. Can be either a simple HTML color string, for example: 'red' or '#00cc00', or an object with the following properties.
\itemize{
\item{"stroke"}{ : string. Default to 'gray'. The color of the chart border, as an HTML color string.}
\item{"strokeWidth"}{ : number. Default to 1. The border width, in pixels.}
\item{"fill"}{ : string. Default to 'white'. The chart fill color, as an HTML color string.}
}}

\item{cameraPosition}{: Object. Default to {horizontal: 1.0, vertical: 0.5, distance: 1.7}. Set the initial rotation and position of the camera. The object cameraPosition contains three parameters: horizontal, vertical, and distance. Parameter horizontal is a value in radians and can have any value (but normally in the range of 0 and 2*Pi). Parameter vertical is a value in radians between 0 and 0.5*Pi. Parameter distance is the (normalized) distance from the camera to the center of the graph, in the range of 0.71 to 5.0. A larger distance puts the graph further away, making it smaller. All parameters are optional.}

\item{height}{: string. Default to '400px'. The height of the graph in pixels or as a percentage.}

\item{keepAspectRatio}{: boolean. Default to true. If keepAspectRatio is true, the x-axis and the y-axis keep their aspect ratio. If false, the axes are scaled such that they both have the same, maximum with.}

\item{showAnimationControls}{: boolean. Default to true. If true, animation controls are created at the bottom of the Graph. The animation controls consists of buttons previous, start/stop, next, and a slider showing the current frame. Only applicable when the provided data contains an animation.}

\item{showGrid}{: boolean. Default to true. If true, grid lines are draw in the x-y surface (the bottom of the 3d graph).}

\item{showPerspective}{: boolean. Default to true. If true, the graph is drawn in perspective: points and lines which are further away are drawn smaller. Note that the graph currently does not support a gray colored bottom side when drawn in perspective.}

\item{showShadow}{: boolean. Default to false. Show shadow on the graph.}

\item{style}{: string. Default to 'dot'. The style of the 3d graph. Available styles: bar, bar-color, bar-size, dot, dot-line, dot-color, dot-size, line, grid, or surface}

\item{tooltip}{: boolean | function. Default to false. Show a tooltip showing the values of the hovered data point. The contents of the tooltip can be customized by providing a callback function as tooltip. In this case the function is called with an object containing parameters x, y, and z argument, and must return a string which may contain HTML.}

\item{valueMax}{: number. Default to none. The maximum value for the value-axis. Only available in combination with the styles dot-color and dot-size.}

\item{valueMin}{: number. Default to none. The minimum value for the value-axis. Only available in combination with the styles dot-color and dot-size.}

\item{verticalRatio}{: number. Default to 0.5. A value between 0.1 and 1.0. This scales the vertical size of the graph When keepAspectRatio is set to false, and verticalRatio is set to 1.0, the graph will be a cube.}

\item{width}{: string. Default to 400px'. The width of the graph in pixels or as a percentage.}

\item{xBarWidth}{: number. Default to none. The width of bars in x direction. By default, the width is equal to the distance between the data points, such that bars adjoin each other. Only applicable for styles 'bar' and 'bar-color'.}

\item{xCenter}{: string. Default to '55\%'. The horizontal center position of the graph, as a percentage or in pixels.}

\item{xMax}{: number. Default to none. The maximum value for the x-axis.}

\item{xMin}{: number. Default to none. The minimum value for the x-axis.}

\item{xStep}{: number. Default to none. Step size for the grid on the x-axis.}

\item{xValueLabel}{: function. Default to none. A function for custom formatting of the labels along the x-axis, for example function (x) {return (x * 100) + '\%'}.}

\item{yBarWidth}{: number. Default to none. The width of bars in y direction. By default, the width is equal to the distance between the data points, such that bars adjoin each other. Only applicable for styles 'bar' and 'bar-color'.}

\item{yCenter}{: string. Default to '45\%'. The vertical center position of the graph, as a percentage or in pixels.}

\item{yMax}{: number. Default to none. The maximum value for the y-axis.}

\item{yMin}{: number. Default to none. The minimum value for the y-axis.}

\item{yStep}{: number. Default to none. Step size for the grid on the y-axis.}

\item{yValueLabel}{: function. Default to none. A function for custom formatting of the labels along the y-axis, for example function (y) {return (y * 100) + '\%'}.}

\item{zMin}{: number. Default to none. The minimum value for the z-axis.}

\item{zMax}{: number. Default to none. The maximum value for the z-axis.}

\item{zStep}{: number. Default to none. Step size for the grid on the z-axis.}

\item{zValueLabel}{: function. Default to none. A function for custom formatting of the labels along the z-axis, for example function (z) {return (z * 100) + '\%'}.}

\item{xLabel}{: String. Default to x. Label on the X axis.}

\item{yLabel}{: String. Default to y. Label on the Y axis.}

\item{zLabel}{: String. Default to z. Label on the Z axis.}

\item{filterLabel}{: String. Default to time. Label for the filter column.}

\item{legendLabel}{: String. Default to value. Label for the style description.}
}
\description{
3D visualization general options. See \url{http://visjs.org/docs/graph3d.html#Configuration_Options}.
}

